requirements.txt
asgiref==3.6.0
azure-storage-blob
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
idna==3.4
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
